creating allScientificValues query param
