+++ search-commons/src/main/java/no/unit/nva/search/resource/ResourceParameter.java
@@ -211,6 +211,7 @@ public enum ResourceParameter implements ParameterKey<ResourceParameter> {
   REFERENCE_EXISTS(CUSTOM, ALL_OF, ENTITY_DESCRIPTION_REFERENCE_CONTEXT_REFERENCE),
   REFERENCED_ID(FUZZY_KEYWORD, ANY_OF, REFERENCE_PUBLICATION_CONTEXT_ID_KEYWORD),
   SCIENTIFIC_VALUE(CUSTOM, ANY_OF, SCIENTIFIC_LEVEL_SEARCH_FIELD),
+  ALL_SCIENTIFIC_VALUES(CUSTOM, ANY_OF, SCIENTIFIC_LEVEL_SEARCH_FIELD),
   //    SCIENTIFIC_VALUE_OF_PARENT(PART_OF, ANY_OF, SCIENTIFIC_VALUE),
   SCIENTIFIC_INDEX_STATUS(KEYWORD, ANY_OF, SCIENTIFIC_INDEX_STATUS_KEYWORD),
   SCIENTIFIC_INDEX_STATUS_NOT(KEYWORD, NOT_ANY_OF, SCIENTIFIC_INDEX_STATUS_KEYWORD),
